requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diskcache==5.6.3
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
llama_cpp_python==0.3.2
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
